+++ src/burn/inc/IBootstrapperApplication.h
@@ -380,6 +380,13 @@ DECLARE_INTERFACE_IID_(IBootstrapperApplication, IUnknown, "53C31D56-49C0-426B-A
     //  IDNOACTION instructs the engine to continue.
     STDMETHOD_(int, OnApplyBegin)() = 0;
 
+    // DEPRECATED: In wix4, this will be merged with OnApplyBegin.
+    // OnApplyNumberOfPhases - called right after OnApplyBegin.
+    //
+    STDMETHOD_(void, OnApplyNumberOfPhases)(
+        __in DWORD dwNumberOfApplyPhases
+        ) = 0;
+
     // OnElevate - called before the engine displays an elevation prompt.
     //             Will only happen once per execution of the engine.
     //